Second Commit

+++ src/good/payment/BitcoinPayment.java
@@ -1,8 +1,10 @@
 package good.payment;
 
+import java.math.BigDecimal;
+
 public class BitcoinPayment extends Payment {
     @Override
-    public void processPayment(double amount) {
+    public void processPayment(BigDecimal amount) {
         System.out.println("Bitcoin ile " + amount + " BTC ödeme yapıldı.");
     }
 }
